+++ client/lib/events.js
