+++ server/public/manifest.json
