--- views/controls.pug
@@ -1,25 +1,0 @@
-html
-  include includes/head.pug
-  +head('Radio DJ Play Controls')
-    script(type='text/javascript', src='/lib/controls.js')
-  
-  body
-    #wrapper
-      h1 Radio DJ Controls
-      #current_info
-        h3
-          | Now Playing: 
-          #current_info_string
-      .playback-controls
-        button(onclick='pause(0)') Play
-        button(onclick='pause(1)') Pause
-        button(onclick='restart()') Restart
-        button(onclick='next()') Next
-        button(onclick='stop()') Stop
-        button(onclick='clear()') Clear Playlist
-      .queue
-        #queue_table
-      .stream-controls
-        select()#stream_selector
-        button(onclick='loadStream()') Load Stream
-
