--- views/event_form.pug
@@ -1,36 +1,0 @@
-html
-  include includes/head.pug
-  +head('Edit Event')
-  body
-    h1= title
-    form#event-form(action=action autocomplete='off')
-      if action=='/event-update'
-        lable ID:
-        input(type='text' name='id' size='4' value=id readonly)
-      lable Event Name:
-      input(type='text' name='name' value=name)
-      lable Event Enabled:
-      input(type='checkbox' name='enabled' checked= enabled=='True')
-      lable Event Category:
-      select(name='catID')
-        each val, i in categories
-          option(value=catIDs[i] selected= i==catID )= val
-      br
-      lable Event Schedule Type:
-      select(name='type')
-        each val, i in types
-          option(value=i selected= i==type )= val
-      lable Event Time:
-      input(type='text' name='time' value=time)
-      lable Event Date:
-      input(type='text' name='date' value=date)
-      lable Event Day:
-      input(type='text' name='day' value=day)
-      lable Event Hours:
-      input(type='text' name='hours' value=hours)
-      br
-      lable Event Actions:
-      br
-      textarea(name='data' rows="10" cols="30")= data
-      br
-      input(type='submit')
