--- views/events.pug
@@ -1,24 +1,0 @@
-html
-  include includes/head.pug
-  +head('Network Events')
-    script(src='/lib/events.js')
-    
-  body
-    #wrapper
-      h1 Radio DJ Events
-      .buttons
-        button(onclick='refreshEvents(drawChart)') Refresh Events
-        button(onclick='editEvent()') Edit Event
-        button(onclick='copyEvent()') Copy Event
-        button(onclick='deleteSelected()') Delete Event
-        button(onclick='loadEvent()') New Event
-      .info
-        #queue_table
-        #event_info
-          #event_ID
-          #event_name
-          #event_category
-          #event_type
-          #event_datetime
-          #event_enabled
-          #event_data
