--- views/history.pug
@@ -1,12 +1,0 @@
-html
-  include includes/head.pug
-  +head('Network Control Panel Home')
-    script(src="/lib/history.js")
-  
-  body
-    #wrapper
-      h1 Radio DJ History
-      .buttons
-        button(onclick='drawChart()') Refresh
-      .info
-        #queue_table
