--- views/index.pug
@@ -1,12 +1,0 @@
-html
-  include includes/head.pug
-  +head('Network Control Panel Home')
-  
-  body
-    h1 Radio DJ Control Panel
-    h3
-      a(href="/controls") Controls
-    h3
-      a(href="/events") Events
-    h3
-      a(href="/history") History
